move logic from AdvancedXboxController into the BaseXboxControllerAdapter, which reduces total number of classes
