+++ src/xbot/common/injection/wpi_factories/RealWPIFactory.java
@@ -23,7 +23,6 @@ import xbot.common.controls.sensors.AdvancedJoystickButton;
 import xbot.common.controls.sensors.AnalogDistanceSensor;
 import xbot.common.controls.sensors.AnalogHIDButton;
 import xbot.common.controls.sensors.DistanceSensor;
-import xbot.common.controls.sensors.AdvancedXboxController;
 import xbot.common.controls.sensors.Lidar;
 import xbot.common.controls.sensors.MockGyro;
 import xbot.common.controls.sensors.RealXboxControllerAdapter;
@@ -36,6 +35,7 @@ import xbot.common.controls.sensors.XJoystick;
 import xbot.common.controls.sensors.XPowerDistributionPanel;
 import xbot.common.controls.sensors.adapters.InertialMeasurementUnitAdapter;
 import xbot.common.controls.sensors.AnalogHIDButton.AnalogHIDDescription;
+import xbot.common.controls.sensors.BaseXboxControllerAdapter;
 import xbot.common.controls.sensors.nav6.Nav6Gyro;
 import xbot.common.controls.sensors.wpi_adapters.AnalogInputWPIAdapater;
 import xbot.common.controls.sensors.wpi_adapters.DigitalInputWPIAdapter;
@@ -193,9 +193,8 @@ public class RealWPIFactory implements WPIFactory {
     }
 
     @Override
-    public AdvancedXboxController getXboxController(int number) {
-        RealXboxControllerAdapter realAdapter = new RealXboxControllerAdapter(number);
-        return new AdvancedXboxController(realAdapter);
+    public BaseXboxControllerAdapter getXboxController(int number) {
+        return new RealXboxControllerAdapter(number);
     }
 
 }
